+++ lib/code_keeper/config.rb
@@ -3,13 +3,12 @@
 module CodeKeeper
   # Provide configuration
   class Config
-    attr_accessor :metrics, :number_of_threads, :format, :metrics_engine
+    attr_accessor :metrics, :number_of_threads, :format
 
     def initialize
       @metrics = %i[cyclomatic_complexity class_length abc_metric]
       @number_of_threads = 2
       @format = :json # json and csv are supported.
-      @metrics_engine = :rubocop_standard
     end
   end
 end