+++ lib/code_keeper/measurement.rb
@@ -15,7 +15,7 @@ module CodeKeeper
       @value = attributes.fetch(:value)
     end
 
-    def legacy_key
+    def score_key
       return scope_name if %i[class module singleton_class].include?(scope_type)
 
       "#{path}:#{scope_name}"