+++ spec/code_keeper_spec.rb
@@ -9,9 +9,5 @@ RSpec.describe CodeKeeper do
     it 'set the default value to metrics' do
       expect(CodeKeeper.configure { |config| config }).to be_a CodeKeeper::Config
     end
-
-    it 'sets rubocop standard engine by default' do
-      expect(CodeKeeper.config.metrics_engine).to eq :rubocop_standard
-    end
   end
 end